Track taps, touches, drags
WHAT WORKS PERFECTLY
I can track touching, dragging, and single finger taps (of any
number: single, double, triple, etc.).

WHAT MOSTLY WORKS
But for counting double-finger double taps (or triple taps or higher)
of two fingers (not one) I had to use a method that seems to only work
most of the time. I get 1.5 or 2.0 as a number for the number of taps
when I double-tap with two fingers, and 2.5 or 3.0 when
I triple tap (with two fingers).

This is stil useful, because I just need to round up the number:
round(1.5) == round(2.0) and round(2.5) == round(3.0).

HOW OFTEN IT WORKS
This works about ... 90% of the time? 95%? 99%?
In any case, so far, I think it's good enough.

TRIPLE FINGER TAPS: OUT OF SCOPE
Triple finger taps don't work so well on my iphone, so... I'm going
to give up on trying to implement them. I don't think it will be
necessary given the current scope of the app.