Replace "button label" with "border label"
WHY
I think it's more clear this way. People kep't trying to press the
button label thinking that it was a button too haha.

NOTE!
I made this commit after making some other changes. I'm not sure
if I was able to seperate out the changes that should go into this
commit from the changes that should go into the next commit...