Add logic to connect add and show buttons-i think
WHAT
Ok, so. Here's what's happening. The "show buttons" need to
be on, if a corresponding "add button" is also on. So the logic
I added here makes that happen. If I press an "add button", the
corresponding "show button" is pressed as well, if its not already
pressed.

WHY
Because if the user/admin is adding an icon to the map, they should
see what other icons of the same type are on the map already, and
they should see the icon that they just added.

NOTE 1
If the user/admin unpresses a "add button", they the corresponding
"show button" should STAY pressed. This is to account for the case
where the user/admin had pressed the "show button" before pressing
the "add button".

NOTE 2
I THINK that this commit consists of just those changes, but
I'm not 100% on that. Still learning how to commit at good times
haha.