Add logic to turn off Show Buttons with Add Button
NOTE
I'm not 100% sure that this commit does only this, but I think it does.
It might be a little mixed with what is supposed to go in the next
commit.

WHAT AND WHY
When the user unpresses a "show button", if a corresponding "add button"
is pressed, it should become unpressed. Because the user
shouldn't be able to add something that they can't see.

E.g. if an "add vending machine" button is pressed, and the user then
unpresses the "show food and drinks" button, then the
"add vending machine" button should become unpressed.