Add creationButton code - chatgpt teachable moment
CONTEXT
I added some code that chatgpt recommended for me to handle
creationButton presses. This code was supposed to make it so that when a
creationButton is pressed, it turns off all other creationButtons, and
turns on the relevant visibilityButton (if the relevant visibilityButton
is off).

But! I realized that the suggestion would make the ToggleButton too
coupled with whether or not it was a visibilityButton or creationButton.
I realized this because of something chatgpt had said about keeping
things uncoupled earlier on.

So I pointed out the possible design issue, and asked it for a solution,
that helps with seperation of concerns. It gave me what I think is a
greeatttt solution! Given my level of coding expertise haha. Definitely
helping me to level up my coding design ability. Making it happen
faster, I think, by showing me good code. Or at least what I can
recognize as good code.

REFACTORING OF OLD CODE
It's a bit of a refactoring of previous code that did this.